refactor: created current monitoring component and changed isLoading initial to true because of glitch
