+++ Client/src/Pages/Monitors/Home/index.jsx
@@ -1,19 +1,17 @@
 import "./index.css";
-import { useEffect, useState } from "react";
+import { useEffect } from "react";
 import { useSelector, useDispatch } from "react-redux";
 import { getUptimeMonitorsByTeamId } from "../../../Features/UptimeMonitors/uptimeMonitorsSlice";
 import { useNavigate } from "react-router-dom";
 import { useTheme } from "@emotion/react";
-import { Box, Button, Stack, Typography } from "@mui/material";
+import { Box, Button, Stack } from "@mui/material";
 import PropTypes from "prop-types";
 import SkeletonLayout from "./skeleton";
 import Fallback from "./fallback";
 import StatusBox from "./StatusBox";
 import Breadcrumbs from "../../../Components/Breadcrumbs";
 import Greeting from "../../../Utils/greeting";
-import MonitorTable from "./MonitorTable";
-import Search from "../../../Components/Inputs/Search";
-import useDebounce from "../../../Utils/debounce";
+import { CurrentMonitoring } from "./CurrentMonitoring";
 
 const Monitors = ({ isAdmin }) => {
 	const theme = useTheme();
@@ -22,19 +20,11 @@ const Monitors = ({ isAdmin }) => {
 	const authState = useSelector((state) => state.auth);
 	const dispatch = useDispatch({});
 
-	//TODO create components, and lower these states.
-	const [search, setSearch] = useState("");
-	const [isSearching, setIsSearching] = useState(false);
-	const debouncedFilter = useDebounce(search, 500);
-
-	const handleSearch = (value) => {
-		setIsSearching(true);
-		setSearch(value);
-	};
-
 	useEffect(() => {
 		dispatch(getUptimeMonitorsByTeamId(authState.authToken));
 	}, [authState.authToken, dispatch]);
+
+	//TODO bring fetching to this component, like on pageSpeed
 
 	const loading = monitorState?.isLoading;
 
@@ -100,57 +90,11 @@ const Monitors = ({ isAdmin }) => {
 									value={monitorState?.monitorsSummary?.monitorCounts?.paused ?? 0}
 								/>
 							</Stack>
-							<Box
-								flex={1}
-								px={theme.spacing(10)}
-								py={theme.spacing(8)}
-								border={1}
-								borderColor={theme.palette.border.light}
-								borderRadius={theme.shape.borderRadius}
-								backgroundColor={theme.palette.background.main}
-							>
-								<Stack
-									direction="row"
-									alignItems="center"
-									mb={theme.spacing(8)}
-								>
-									<Typography
-										component="h2"
-										variant="h2"
-										fontWeight={500}
-										letterSpacing={-0.2}
-									>
-										Actively monitoring
-									</Typography>
-									<Box
-										className="current-monitors-counter"
-										color={theme.palette.text.primary}
-										border={1}
-										borderColor={theme.palette.border.light}
-										backgroundColor={theme.palette.background.accent}
-									>
-										{totalMonitors}
-									</Box>
-									<Box
-										width="25%"
-										minWidth={150}
-										ml="auto"
-									>
-										<Search
-											options={monitorState.monitorsSummary.monitors}
-											filteredBy="name"
-											inputValue={search}
-											handleInputChange={handleSearch}
-										/>
-									</Box>
-								</Stack>
-								<MonitorTable
-									isAdmin={isAdmin}
-									filter={debouncedFilter}
-									setLoading={setIsSearching}
-									isSearching={isSearching}
-								/>
-							</Box>
+							<CurrentMonitoring
+								isAdmin={isAdmin}
+								monitors={monitorState.monitorsSummary.monitors}
+								totalMonitors={totalMonitors}
+							/>
 						</>
 					)}
 				</>
